+++ common/variables.go
@@ -298,6 +298,10 @@ func CaBundleFile() string {
 	return ExpandPath(filepath.Join(RobocorpHome(), "ca-bundle.pem"))
 }
 
+func CaBundleDir() string {
+	return ExpandPath(RobocorpHome())
+}
+
 func DefineVerbosity(silent, debug, trace bool) {
 	override := os.Getenv(RCC_VERBOSITY)
 	switch {